Added car, event handler, collision detector and much more

+++ Frogger/src/UserInterface.h
@@ -2,9 +2,8 @@
 #include <stdio.h>
 #include <cstring>
 
-#include <SDL_ttf.h>
+#include "Global.h"
 #include "TextureManager.h"
-#include "Game.h"
 
 
 class UserInterface
@@ -14,7 +13,7 @@ public:
 	~UserInterface();
 
 	void init(int surface_height, int surface_width, int window_hight);
-	void update_info(double world_time, int fps);
+	void update_info(double world_time, double fps);
 	void render();
 	SDL_Texture* get_texture();
 	SDL_Texture* get_texture_text();
